requirements.txt
python-docx
requests
beautifulsoup4
pdfplumber
pillow
pytesseract
fpdf
